replaced strtok_r dependency with homemade string tokenizer
This was done after it was determined that the strtok_r function
would cause Sporth to crash when doing on-the-fly coding.
The replacement function is much simpler and a less robust,
but it does the whitespace parsing needed for gen_vals.

strtok_r also raises a warning when in c99 mode, so perhaps its
for the best?